+++ i3configger/cli.py
@@ -43,9 +43,8 @@ def _parse_args(p):
     p.add_argument('--i3-refresh-msg', action="store", default='reload',
                    choices=['restart', 'reload', 'nop'],
                    help="i3-msg to send after build")
-    # TODO automatically deactivate notify if no one will pick them up
-    p.add_argument('--no-notify', action="store_true", default=False,
-                   help="deactivate notification via notify-send")
+    p.add_argument('--notify', action="store_true", default=False,
+                   help="show build notification via notify-send")
     p.add_argument('--log', action="store", default=None,
                    help="i3configgerPath to where log should be stored")
     p.add_argument('--load-config', action="store", default=None,