+++ i3configger/context.py
@@ -1,21 +1,21 @@
 import logging
 import typing as t
+from string import Template
 
 from i3configger import base, exc, partials
 
 log = logging.getLogger(__name__)
 
 
-def create(prts: t.Iterable[partials.Partial]) -> dict:
-    """Read variables from all partials into a dictionary and resolve them."""
+def merge(item: t.Iterable[t.Union[dict, partials.Partial]]) -> dict:
+    """Take items, merge contexts, resolve and clean them variables"""
     ctx = {}
-    for context in [fetch(p.payload) for p in prts]:
-        ctx.update(context)
-    ctx = resolve(ctx)
-    assert all(k.startswith(base.VAR_MARK) for k in ctx.keys()), ctx
+    for elem in item:
+        if not isinstance(elem, dict):
+            elem = elem.context
+        ctx.update(elem)
+    ctx = resolve_variables(ctx)
     ctx = remove_variable_markers(ctx)
-    assert not any(k.startswith(base.VAR_MARK) for k in ctx.keys()), ctx
-    assert not any(v.startswith(base.VAR_MARK) for v in ctx.values()), ctx
     return ctx
 
 
@@ -28,16 +28,7 @@ def remove_variable_markers(ctx: dict) -> dict:
     return cleaned
 
 
-def enhance(ctx, otherContexts):
-    eCtx = dict(ctx)
-    for oc in otherContexts:
-        if isinstance(oc, partials.Partial):
-            oc = create([oc])
-        eCtx.update(oc)
-    return eCtx
-
-
-def resolve(context: dict) -> dict:
+def resolve_variables(context: dict) -> dict:
     """If variables are set by a variable, replace them by their value."""
     if not any(v.startswith(base.VAR_MARK) for v in context.values()):
         log.debug("needs no resolution: %s", context)
@@ -58,13 +49,10 @@ def resolve(context: dict) -> dict:
     return resolvedContext
 
 
-def fetch(content: str) -> dict:
-    ctx = {}
-    for line in content.splitlines():
-        line = line.strip()
-        if not line.startswith(base.SET_MARK):
-            continue
-        payload = line.split(maxsplit=1)[1]
-        key, value = payload.split(maxsplit=1)
-        ctx[key] = value
-    return ctx
+def substitute(content, ctx):
+    """Substitute all variables with their values.
+
+    Works out of the box, because '$' is the standard substitution
+    marker for string.Template
+    """
+    return Template(content).safe_substitute(ctx)